QualityGate optimization
